+++ src/lib.rs
@@ -24,8 +24,6 @@ pub mod i2c;
 #[cfg(feature = "device-selected")]
 pub mod prelude;
 #[cfg(feature = "device-selected")]
-pub mod pwm;
-#[cfg(feature = "device-selected")]
 pub mod rcc;
 #[cfg(any(feature = "py32f003", feature = "py32f030"))]
 pub mod rtc;