+++ src/timer/hal_1.rs
@@ -5,7 +5,8 @@
 
 use embedded_hal::delay::DelayNs;
 
-use super::{Delay, Instance, SysDelay};
+use super::{Delay, Instance, OcPin, PwmChannel, SysDelay, WithPwm};
+use core::convert::Infallible;
 use fugit::ExtU32Ceil;
 
 impl DelayNs for SysDelay {
@@ -32,16 +33,16 @@ impl<TIM: Instance, const FREQ: u32> DelayNs for Delay<TIM, FREQ> {
     }
 }
 
-// impl<TIM: Instance + WithPwm, const C: u8> embedded_hal::pwm::ErrorType for PwmChannel<TIM, C> {
-//     type Error = Infallible;
-// }
+impl<TIM: Instance + WithPwm, CH: OcPin> embedded_hal::pwm::ErrorType for PwmChannel<TIM, CH> {
+    type Error = Infallible;
+}
 
-// impl<TIM: Instance + WithPwm, const C: u8> embedded_hal::pwm::SetDutyCycle for PwmChannel<TIM, C> {
-//     fn max_duty_cycle(&self) -> u16 {
-//         self.get_max_duty()
-//     }
-//     fn set_duty_cycle(&mut self, duty: u16) -> Result<(), Self::Error> {
-//         self.set_duty(duty);
-//         Ok(())
-//     }
-// }
+impl<TIM: Instance + WithPwm, CH: OcPin> embedded_hal::pwm::SetDutyCycle for PwmChannel<TIM, CH> {
+    fn max_duty_cycle(&self) -> u16 {
+        self.get_max_duty()
+    }
+    fn set_duty_cycle(&mut self, duty: u16) -> Result<(), Self::Error> {
+        self.set_duty(duty);
+        Ok(())
+    }
+}